fix filter at reading everywhere

+++ internal/entity/link_params.go
@@ -77,8 +77,8 @@ func (f *TestSkillFilter) Validate() error {
 type UserSkillFilter struct {
 	UF               UserFilter
 	SF               SkillFilter
-	ProficiencyLevel []int `json:"proficiency_level,omitempty"`
-	Limit            uint  `json:"limit,omitempty"`
+	ProficiencyLevel []int64 `json:"proficiency_level,omitempty"`
+	Limit            uint    `json:"limit,omitempty"`
 }
 
 func (f *UserSkillFilter) Validate() error {
